refactor: `add` and `delete` in base class
